+++ components/atoms/category.tsx
@@ -1,7 +1,7 @@
 export default function Category({ category }: { category: any }) {
   return (
     <a
-      href={`/category/${category.fields.slug}`}
+      href={`/categories/${category.fields.slug}`}
       className="inline-flex items-center px-2 py-3 m-1 text-sm font-bold text-gray-800 bg-gray-100 rounded-sm hover:bg-gray-200"
       style={{
         backgroundColor: category.fields.backgroundColor,
