+++ components/atoms/tag.tsx
@@ -1,7 +1,7 @@
 export default function Tag({ tag }: { tag: any }) {
   return (
     <a
-      href={`/tag/${tag.fields.slug}`}
+      href={`/tags/${tag.fields.slug}`}
       className="inline-flex items-center px-2 py-3 m-1 text-sm font-bold text-gray-800 bg-gray-100 rounded-sm hover:bg-gray-200"
       style={{
         backgroundColor: tag.fields.backgroundColor,
