+++ pages/categories/[slug].tsx
@@ -8,13 +8,13 @@ import client from '@/lib/contentful';
 import { PER_PAGE, CONTENT_TYPE, BASE_URL } from '@/lib/constants';
 import { BreadcrumbPage } from '@/@types/index';
 
-function IndexPage({ initialArticles, total, tag, pages }: any) {
+function IndexPage({ initialArticles, total, category, pages }: any) {
   const [articles, setArticles] = useState(initialArticles);
 
   const getArticles = async (page: number) => {
     const res = await client.getEntries({
       content_type: CONTENT_TYPE.ARTICLE,
-      'fields.tag.sys.id': tag.sys.id,
+      'fields.type.sys.id': category.sys.id,
       order: '-sys.updatedAt',
       limit: PER_PAGE,
       skip: PER_PAGE * (page - 1),
@@ -26,11 +26,15 @@ function IndexPage({ initialArticles, total, tag, pages }: any) {
   return (
     <>
       <Head>
-        <title>{tag.fields.name}</title>
-        <meta property="og:title" content={tag.fields.name} key="og_title" />
+        <title>{category.fields.name}</title>
+        <meta
+          property="og:title"
+          content={category.fields.name}
+          key="og_title"
+        />
         <meta
           property="og:url"
-          content={`${BASE_URL}/tag/${tag.fields.slug}`}
+          content={`${BASE_URL}/tags/${category.fields.slug}`}
           key="og_url"
         />
       </Head>
@@ -43,7 +47,7 @@ function IndexPage({ initialArticles, total, tag, pages }: any) {
           hasMore={articles.length < total}
           loader={
             <div className="mx-5 my-2 lg:mx-auto" key={1}>
-              Loading ...
+              ロード中 ...
             </div>
           }
           useWindow={true}
@@ -58,23 +62,23 @@ function IndexPage({ initialArticles, total, tag, pages }: any) {
 }
 
 export const getStaticPaths: GetStaticPaths = async () => {
-  let tags = await client.getEntries({
-    content_type: CONTENT_TYPE.TAG,
+  let categories = await client.getEntries({
+    content_type: CONTENT_TYPE.CATEGORY,
     limit: 1,
   });
-  const total = tags.total;
+  const total = categories.total;
   const maxPage = Math.ceil(total / PER_PAGE);
   const maxPageArray = [...Array(maxPage).keys()].map((i) => ++i);
 
   const paths = [];
   for (const l of maxPageArray) {
-    tags = await client.getEntries({
-      content_type: CONTENT_TYPE.TAG,
+    categories = await client.getEntries({
+      content_type: CONTENT_TYPE.CATEGORY,
       limit: PER_PAGE,
       skip: PER_PAGE * (l - 1),
     });
-    for (const tag of tags.items) {
-      paths.push({ params: { slug: tag.fields.slug } });
+    for (const category of categories.items) {
+      paths.push({ params: { slug: category.fields.slug } });
     }
   }
 
@@ -82,9 +86,9 @@ export const getStaticPaths: GetStaticPaths = async () => {
 };
 
 export const getStaticProps: GetStaticProps = async ({ params }) => {
-  const tag = await client
+  const category = await client
     .getEntries({
-      content_type: CONTENT_TYPE.TAG,
+      content_type: CONTENT_TYPE.CATEGORY,
       'fields.slug': params?.slug,
     })
     .then((res: any) => res.items[0])
@@ -92,20 +96,15 @@ export const getStaticProps: GetStaticProps = async ({ params }) => {
 
   const articles = await client.getEntries({
     content_type: CONTENT_TYPE.ARTICLE,
-    'fields.tag.sys.id': tag.sys.id,
+    'fields.type.sys.id': category.sys.id,
     order: '-sys.updatedAt',
     limit: PER_PAGE,
   });
 
   const pages: BreadcrumbPage[] = [
     {
-      name: tag.fields.type.fields.name,
-      href: `/category/${tag.fields.type.fields.slug}`,
-      current: false,
-    },
-    {
-      name: tag.fields.name,
-      href: `/tag/${tag.fields.slug}`,
+      name: category.fields.name,
+      href: `/categories/${category.fields.slug}`,
       current: true,
     },
   ];
@@ -114,7 +113,7 @@ export const getStaticProps: GetStaticProps = async ({ params }) => {
     props: {
       initialArticles: articles.items,
       total: articles.total,
-      tag: tag,
+      category: category,
       pages: pages,
     },
   };
